src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/custom-pseudo-scope-expected.txt

CONSOLE ERROR: line 10: Uncaught TypeError: host1.createShadowRoot is not a function
